requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
sqlalchemy
mysql-connector-python  
jupyterlab              
